Initial commit. July 2023. Standing on the shoulders of giants. This is a Sonos remote control for a Lilygo T Display S3. (Esp32 based)
Some things to be aware of:
Use the builtin LittleFS, not an external library.
Use Lilygo's TFT_eSPI library (in the lib folder) NOT one from the Platformio library manager.

Upload the files in the data folder to the initialized SPIFFs of the ESP32.
Use the platformio upload file system task.